alp 需求修改 更换文字，增加notes列

+++ OASys-Frontend-master/src/components/ImportData.vue
@@ -58,6 +58,9 @@
                         <el-button type="primary" @click="handleGetGoods()">search</el-button>
                     </el-col>
                     <el-col :span="1.5">
+                        <el-button type="primary" @click="handleGetWarningGoods()">{{warningContent}}</el-button>
+                    </el-col>
+                    <el-col :span="1.5">
                         <el-button type="primary" @click="handleExportGoods()">export</el-button>
                     </el-col>
                     <el-col :span="1.5">
@@ -66,11 +69,11 @@
                 </el-row>
             </el-form>
         </el-card>
-        <el-table border :data="goodsList" ref="multipleTable"
+        <el-table border :data="goodsList" ref="multipleTable" :row-style="changeColor"
                   style="width: 100%">
             <el-table-column
-                    prop="id"
-                    label="id"
+                    prop="rank"
+                    label="rank"
                     align="center"
                     width="100">
             </el-table-column>
@@ -101,7 +104,6 @@
                     align="center">
             </el-table-column>
             <el-table-column
-                    :show-overflow-tooltip="true"
                     prop="description"
                     label="Description"
                     align="center">
@@ -122,13 +124,8 @@
                     align="center">
             </el-table-column>
             <el-table-column
-                    prop="annualStock"
-                    label="annual stock"
-                    align="center">
-            </el-table-column>
-            <el-table-column
                     prop="autoReplenishRate"
-                    label="auto replenish rate"
+                    label="Safety Stock Line"
                     align="center">
             </el-table-column>
             <el-table-column
@@ -136,11 +133,15 @@
                     label="lead time"
                     align="center">
             </el-table-column>
-
+            <el-table-column
+                    prop="notes"
+                    label="Notes"
+                    align="center">
+            </el-table-column>
             <el-table-column
                     label="操作"
                     align="center"
-                    width="200"
+                    width="300"
                     fixed="right">
                 <template slot-scope="scope">
                     <el-button type="primary" size="mini"
@@ -148,6 +149,9 @@
                     </el-button>
                     <el-button type="danger" size="mini"
                                @click="deleteGoods(scope.row)">del
+                    </el-button>
+                    <el-button type="info" size="mini"
+                               @click="ordering(scope.row)">ordering
                     </el-button>
                 </template>
             </el-table-column>
@@ -166,7 +170,7 @@
             </el-pagination>
         </div>
         <el-dialog :visible.sync="editDialogVisible" :rules="rules" :ref="goodsToEdit">
-            <el-form label-width="80px" v-model="goodsToEdit">
+            <el-form label-width="80px" :model="goodsToEdit">
                 <el-input type="hidden" v-model="goodsToEdit.id"/>
                 <el-form-item label="tagid" prop="tagid">
                     <el-input v-model="goodsToEdit.tagid"/>
@@ -184,25 +188,31 @@
                     <el-input v-model="goodsToEdit.manufacturerPartNumber"/>
                 </el-form-item>
                 <el-form-item label="description" prop="description">
-                    <el-input type="textarea"  v-model="goodsToEdit.description"/>
+                    <el-input type="textarea" v-model="goodsToEdit.description"/>
                 </el-form-item>
                 <el-form-item label="date" prop="date">
-                    <el-input v-model="goodsToEdit.date"/>
+                    <el-date-picker
+                            v-model="goodsToEdit.date"
+                            type="datetime"
+                            format="yyyy-MM-dd HH:mm:ss"
+                            value-format="yyyy-MM-dd HH:mm:ss"
+                            placeholder="choose date">
+                    </el-date-picker>
                 </el-form-item>
                 <el-form-item label="stockQty" prop="stockQty">
-                    <el-input v-model="goodsToEdit.stockQty"/>
+                    <el-input type="number" v-model="goodsToEdit.stockQty"/>
                 </el-form-item>
                 <el-form-item label="price" prop="price">
                     <el-input type="number" v-model="goodsToEdit.price"/>
                 </el-form-item>
-                <el-form-item label="annualStock" prop="annualStock">
-                    <el-input v-model="goodsToEdit.annualStock"/>
-                </el-form-item>
                 <el-form-item label="autoReplenishRate" prop="autoReplenishRate">
-                    <el-input v-model="goodsToEdit.autoReplenishRate"/>
+                    <el-input type="number" v-model="goodsToEdit.autoReplenishRate"/>
                 </el-form-item>
                 <el-form-item label="leadTime" prop="leadTime">
                     <el-input v-model="goodsToEdit.leadTime"/>
+                </el-form-item>
+                <el-form-item label="Notes" prop="notes">
+                    <el-input v-model="goodsToEdit.notes"/>
                 </el-form-item>
             </el-form>
             <div class="dialog-footer" slot="footer">
@@ -216,9 +226,9 @@
         </el-dialog>
 
         <el-dialog :visible.sync="addDialogVisible">
-            <el-form v-model="goodsToAdd" :rules="rules" :ref="goodsToAdd">
+            <el-form :model="goodsToAdd" :rules="rules" :ref="goodsToAdd">
                 <el-form-item label="tagid" prop="tagid">
-                    <el-input v-model="goodsToAdd.tagid"/>
+                    <el-input v-model.number="goodsToAdd.tagid"/>
                 </el-form-item>
                 <el-form-item label="Component type" prop="componentType">
                     <el-input v-model="goodsToAdd.componentType"/>
@@ -236,7 +246,13 @@
                     <el-input type="textarea" v-model="goodsToAdd.description"/>
                 </el-form-item>
                 <el-form-item label="date" prop="date">
-                    <el-input v-model="goodsToAdd.date"/>
+                    <el-date-picker
+                            v-model="goodsToAdd.date"
+                            type="datetime"
+                            format="yyyy-MM-dd HH:mm:ss"
+                            value-format="yyyy-MM-dd HH:mm:ss"
+                            placeholder="choose date">
+                    </el-date-picker>
                 </el-form-item>
                 <el-form-item label="stockQty" prop="stockQty">
                     <el-input v-model="goodsToAdd.stockQty"/>
@@ -244,14 +260,14 @@
                 <el-form-item label="price" prop="price">
                     <el-input v-model="goodsToAdd.price"/>
                 </el-form-item>
-                <el-form-item label="annualStock" prop="annualStock">
-                    <el-input v-model="goodsToEdit.annualStock"/>
-                </el-form-item>
                 <el-form-item label="autoReplenishRate" prop="autoReplenishRate">
-                    <el-input v-model="goodsToEdit.autoReplenishRate"/>
+                    <el-input type="number" v-model="goodsToAdd.autoReplenishRate"/>
                 </el-form-item>
                 <el-form-item label="leadTime" prop="leadTime">
-                    <el-input v-model="goodsToEdit.leadTime"/>
+                    <el-input v-model="goodsToAdd.leadTime"/>
+                </el-form-item>
+                <el-form-item label="Notes" prop="notes">
+                    <el-input v-model="goodsToAdd.notes"/>
                 </el-form-item>
             </el-form>
             <div class="dialog-footer" slot="footer">
@@ -263,10 +279,43 @@
                 </el-button>
             </div>
         </el-dialog>
+        <el-dialog :visible.sync="orderingDialogVisible">
+            <h1>ordering</h1>
+            <el-form label-width="160px" :model="order" :rules="rules" :ref="order">
+                <el-form-item label="Item No" prop="itemNo">
+                    <el-input type="number" v-model="order.itemNo"/>
+                </el-form-item>
+                <el-form-item label="Qty shipped" prop="qtyShipped">
+                    <el-input type="number" v-model.number="order.qtyShipped"/>
+                </el-form-item>
+                <el-form-item label="Customer PO#">
+                    <el-input v-model="order.customerPo"/>
+                </el-form-item>
+                <el-form-item label="Project#">
+                    <el-input v-model="order.project"/>
+                </el-form-item>
+                <el-form-item label="Shipment date">
+                    <el-date-picker
+                            v-model="order.shipmentDate"
+                            type="datetime"
+                            value-format="yyyy-MM-dd HH:mm:ss"
+                            placeholder="选择日期时间">
+                    </el-date-picker>
+                </el-form-item>
+            </el-form>
+            <div class="dialog-footer" slot="footer">
+                <el-button @click="orderingDialogVisible = false">
+                    cancel
+                </el-button>
+                <el-button @click="updateOrdering" type="primary">
+                    ok
+                </el-button>
+            </div>
+        </el-dialog>
         <el-dialog :visible.sync="uploadFileDialog" width="26%">
             <el-upload ref="upload"
                        :file-list="flist"
-                    drag action="/api/importData" multiple :data={}
+                       drag action="/api/importData" multiple :data={}
                        :on-success="uploadSuccess">
                 <i class="el-icon-upload"></i>
                 <div class="el-upload__text">pull file here,or<em>click upload</em></div>
@@ -276,31 +325,45 @@
 </template>
 
 <script>
-    import {addGoods, getGoods, updateGoods, deleteGoods,exportGoods,deleteAllGoods} from "../utils/api";
+    import {
+        addGoods,
+        getGoods,
+        updateGoods,
+        deleteGoods,
+        exportGoods,
+        deleteAllGoods,
+        getWarningGoods,
+        addOrdering
+    } from "../utils/api";
     import {mapState} from "vuex";
 
     export default {
         name: "ImportData",
         data() {
             return {
-                flist:[],
+                warningContent:'warning',
+                warning:false,
+                flist: [],
                 goodsList: [],
                 goodsToEdit: {},
                 goodsToAdd: {},
+                goodsToOrdering: {},
+                order: {},
                 total: 0,
                 addFolderDialog: false,
                 uploadFileDialog: false,
                 renameFileDialog: false,
                 editDialogVisible: false,
+                orderingDialogVisible: false,
                 addDialogVisible: false,
-                currentPage4:1,
+                currentPage4: 1,
                 searchParams: {},
                 rules: {
                     tagid: [
+                        {required: true, message: 'required', trigger: "blur"},
                         {
-                            type: "number",
-                            required: true,
-                            message: "input number tagid",
+                            type: 'number',
+                            message: "number required",
                             trigger: "blur"
                         }
                     ],
@@ -346,6 +409,16 @@
                             trigger: "blur"
                         }
                     ],
+                    qtyShipped: [
+                        {
+                            required: true, message: 'required', trigger: "blur"
+                        },
+                        {
+                            type: "number",
+                            message: "required",
+                            trigger: "blur"
+                        }
+                    ],
                 },
             }
         },
@@ -358,6 +431,15 @@
             ...mapState(["auth"])
         },
         methods: {
+            changeColor(row){
+                let colors = ['#CCFF33','#CCFF99','#66CC33','#99FFFF','#CCFFFF','#FFCCCC','#FFFFFF','white']
+                var r = row.row.rank
+                window.console.log(JSON.stringify(row))
+                let i = r % 7
+                let color = colors[i]
+
+                return { "background-color": color }
+            },
             handleAddGoods() {
                 this.addDialogVisible = true;
             },
@@ -366,7 +448,7 @@
                     if (response && response.status === "success") {
                         this.$message.success(response.message)
                         this.addDialogVisible = false
-                        this.goodsToAdd={}
+                        this.goodsToAdd = {}
                         this.handleGetGoods()
                     }
                 })
@@ -374,6 +456,10 @@
             handleEdit(row) {
                 this.goodsToEdit = row
                 this.editDialogVisible = true;
+            },
+            ordering(row) {
+                this.goodsToOrdering = row
+                this.orderingDialogVisible = true;
             },
             deleteGoods(row) {
                 this.$confirm("delete , continue?")
@@ -388,31 +474,64 @@
                     })
             },
             handleGetGoods() {
-                getGoods(this.searchParams).then(response => {
-                    if (response && response.status === "success") {
-                        this.total = response.total
-                        this.goodsList = response.object
-                    }
-                })
+                if (this.warning === true){
+                    getWarningGoods(this.searchParams).then(response => {
+                        if (response && response.status === "success") {
+                            this.total = response.total
+                            this.goodsList = response.object
+                        }
+                    })
+                }else{
+                    getGoods(this.searchParams).then(response => {
+                        if (response && response.status === "success") {
+                            this.total = response.total
+                            this.goodsList = response.object
+                        }
+                    })
+                }
+            },
+            handleGetWarningGoods() {
+                this.warning = !this.warning;
+                if (this.warning){
+                    this.warningContent = 'cancel warning'
+                }else{
+                    this.warningContent = 'warning'
+                }
+                this.handleGetGoods();
             },
             handleExportGoods() {
-                exportGoods(this.searchParams,'data.xls');
+                exportGoods(this.searchParams, 'data.xls');
             },
             handleDeleteAllGoods() {
-                deleteAllGoods().then(response=>{
-                    if (response && response.status === "success") {
-                        this.$message.success(response.message)
-                        this.handleGetGoods()
-                    }else{
-                        this.$message.error(response.message)
-                    }
-                });
+                this.$confirm("delete , continue?")
+                    .then(() => {
+                        deleteAllGoods().then(response => {
+                            if (response && response.status === "success") {
+                                this.$message.success(response.message)
+                                this.handleGetGoods()
+                            } else {
+                                this.$message.error(response.message)
+                            }
+                        });
+                    });
             },
             updateGoods() {
                 updateGoods(this.goodsToEdit).then(response => {
                     if (response && response.status === "success") {
                         this.$message.success(response.message)
                         this.editDialogVisible = false
+                        this.handleGetGoods()
+                    }
+                })
+            },
+            updateOrdering() {
+                this.order.goodsPartNumber = this.goodsToOrdering.manufacturerPartNumber
+                this.order.goodsId = this.goodsToOrdering.id
+                addOrdering(this.order).then(response => {
+                    if (response && response.status === "success") {
+                        this.$message.success(response.message)
+                        this.orderingDialogVisible = false
+                        this.order.qtyShipped = 0
                         this.handleGetGoods()
                     }
                 })
@@ -426,7 +545,7 @@
                 this.uploadFileDialog = false
                 this.$refs.upload.clearFiles()
             },
-            handlePageSizeChange(pageSize){
+            handlePageSizeChange(pageSize) {
                 this.searchParams.pageSize = pageSize;
                 this.handleGetGoods();
             },
